commit to saving code

+++ EmployeeManager/src/main/java/com/example/EmployeeManager/Model/Employee/DeleteEmployee.java
@@ -1,7 +1,10 @@
 package com.example.EmployeeManager.Model.Employee;
 
 import com.example.EmployeeManager.Entity.Employee;
+import com.example.EmployeeManager.Entity.Team;
+import com.example.EmployeeManager.HandleException.DependenceDataException;
 import com.example.EmployeeManager.Repository.EmployeeRepository;
+import com.example.EmployeeManager.Repository.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +17,9 @@ public class DeleteEmployee {
     @Autowired
     private EmployeeRepository employeeRepository;
 
+    @Autowired
+    private TeamRepository teamRepository;
+
     private List<Integer> idList;
 
     public void setIdList(List<Integer> idList) {
@@ -23,6 +29,12 @@ public class DeleteEmployee {
     public void deleteEmployeeById() {
         for (Integer id : idList) {
             Optional<Employee> optionalEmployee = employeeRepository.findById(id);
+
+            Optional<Team> optionalTeam = teamRepository.findByManagerId(id);
+            if (optionalTeam.isPresent()) {
+                throw new DependenceDataException("employee is a manager");
+            }
+
             if (optionalEmployee.isPresent()) {
                 employeeRepository.deleteById(id);
             }
